automated_respiration_analysis/requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.0
kiwisolver==1.4.9
matplotlib==3.10.6
mysql-connector-python==9.4.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==78.1.1
six==1.17.0
tzdata==2025.2
wheel==0.45.1
